Split BulkSyncInterface into 3 implementations
